% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSH4GWASCatalog.R
\name{generateSH}
\alias{generateSH}
\title{Generate a shell script for parallel donwloading given a vector of urls}
\usage{
generateSH(urls, outPath)
}
\arguments{
\item{urls}{A vector of urls}

\item{outPath}{Path for saving the shell script}
}
\description{
Generate a shell script for parallel donwloading given a vector of urls
}
