% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSH4GWASCatalog.R
\name{getHarmonizedUrlsFromGWASCatalog}
\alias{getHarmonizedUrlsFromGWASCatalog}
\title{Get harmonized urls given a accession number}
\usage{
getHarmonizedUrlsFromGWASCatalog(accessionNumberList)
}
\arguments{
\item{accessionNumberList}{A vector contains accession number lists}
}
\description{
Get harmonized urls given a accession number
}
