% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSH4GWASCatalog.R
\name{getSH4GWASCatalog}
\alias{getSH4GWASCatalog}
\title{Generate a shell script for downloading summary-level data from GWAS Catalog}
\usage{
getSH4GWASCatalog(accessionNumberList)
}
\arguments{
\item{accessionNumberList}{The accession number or a vector of accession numbers}
}
\description{
Generate a shell script for downloading summary-level data from GWAS Catalog
}
