% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSH4GWASCatalog.R
\name{getSH4HarmonizedGWASCatalog}
\alias{getSH4HarmonizedGWASCatalog}
\title{Generate a shell script for parallel downloading harmonised summary-level data from GWAS Catalog}
\usage{
getSH4HarmonizedGWASCatalog(accessionNumberList)
}
\arguments{
\item{accessionNumberList}{The accession number or a vector of accession numbers}
}
\description{
Generate a shell script for parallel downloading harmonised summary-level data from GWAS Catalog
}
