% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSH4GWASCatalog.R
\name{getUrlFromGWASCatalog}
\alias{getUrlFromGWASCatalog}
\title{Get unharmonised urls from GWAS Catalog}
\usage{
getUrlFromGWASCatalog(accessionNumber)
}
\arguments{
\item{accessionNumber}{An accession number}
}
\description{
Get unharmonised urls from GWAS Catalog
}
