% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSH4GWASCatalog.R
\name{getHarmonizedUrlsFromGWASCatalog}
\alias{getHarmonizedUrlsFromGWASCatalog}
\title{Get harmonised urls given a accession number}
\usage{
getHarmonizedUrlsFromGWASCatalog(accessionNumberLists)
}
\arguments{
\item{accessionNumberLists}{A vector contains accession number lists}
}
\description{
Get harmonised urls given a accession number
}
